requirements.txt
Django==4.2
django-autoslug==1.9.9
django-cors-headers==4.6.0
djangorestframework==3.14.0
djoser==2.3.1
drf-yasg==1.21.8
pillow==11.1.0
psycopg2-binary==2.9.10
django-heroku==0.3.1
gunicorn==23.0.0
